new peer command changes
These changes are purely for the chia peer command
