+++ chia/cmds/chia.py
@@ -8,6 +8,7 @@ from chia.cmds.init import init_cmd
 from chia.cmds.keys import keys_cmd
 from chia.cmds.netspace import netspace_cmd
 from chia.cmds.passphrase import passphrase_cmd
+from chia.cmds.peer import peer_cmd
 from chia.cmds.plots import plots_cmd
 from chia.cmds.rpc import rpc_cmd
 from chia.cmds.show import show_cmd
@@ -132,6 +133,7 @@ cli.add_command(netspace_cmd)
 cli.add_command(farm_cmd)
 cli.add_command(plotters_cmd)
 cli.add_command(db_cmd)
+cli.add_command(peer_cmd)
 cli.add_command(passphrase_cmd)
